Rename the packages and add some features to PomodoroModel
