+++ app/src/main/AndroidManifest.xml
@@ -20,6 +20,6 @@
         </activity>
         <activity android:name="com.example.course_project_tot.CalendarActivity"/>
         <activity android:name="com.example.course_project_tot.CreateGoalActivity"/>
-        <activity android:name="com.example.course_project_tot.Ucontroller.PomodoroController"/>
+        <activity android:name="com.example.course_project_tot.Activity.PomodoroActivity"/>
     </application>
 </manifest>
