+++ app/src/main/java/com/example/course_project_tot/Controller/ILoginController.java
@@ -1,4 +1,4 @@
-package com.example.course_project_tot.Ucontroller;
+package com.example.course_project_tot.Controller;
 
 public interface ILoginController {
     void OnLogin(String email,String Password);
