+++ app/src/main/java/com/example/course_project_tot/CreateGoalPresenter.java
@@ -1,7 +1,7 @@
 package com.example.course_project_tot;
 
 import android.widget.EditText;
-import com.example.course_project_tot.Ucontroller.LoginController;
+import com.example.course_project_tot.Controller.LoginController;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
