+++ app/src/main/java/com/example/course_project_tot/Modele/IUser.java
@@ -1,4 +1,4 @@
-package com.example.course_project_tot.UModel;
+package com.example.course_project_tot.Modele;
 
 public interface IUser {
     String getEmail();
