+++ app/src/main/java/com/example/course_project_tot/Modele/User.java
@@ -1,4 +1,4 @@
-package com.example.course_project_tot.UModel;
+package com.example.course_project_tot.Modele;
 
 import android.text.TextUtils;
 import android.util.Patterns;
