+++ app/src/main/java/com/example/course_project_tot/View/ILoginView.java
@@ -1,4 +1,4 @@
-package com.example.course_project_tot.UView;
+package com.example.course_project_tot.View;
 
 public interface ILoginView {
     void OnLoginSuccess(String message);
