+++ app/src/main/res/layout/activity_pomodoro.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Ucontroller.PomodoroController">
+    tools:context=".Activity.PomodoroActivity">
 
     <TextView
         android:id="@+id/textView_view_countdown"
